boiler plate of the App class
