+++ src/main/scala/nl.knaw.dans.easy.solr4files/Command.scala
@@ -49,8 +49,8 @@ object Command extends App with DebugEnhancedLogging {
       .collect {
         case update @ commandLine.update => app.update(update.baseUri())
         case delete @ commandLine.delete => app.delete(delete.baseUri())
-        case init @ commandLine.init if init.bagStore.isSupplied => app.init(init.uri())
-        case init @ commandLine.init if !init.bagStore.isSupplied => app.initAll(init.uri())
+        case init @ commandLine.init if init.bagStore.isSupplied => app.initSingleStore(init.uri())
+        case init @ commandLine.init if !init.bagStore.isSupplied => app.initAllStores(init.uri())
         case commandLine.runService => ???
       }
       .getOrElse(Failure(new IllegalArgumentException(s"Unknown command: ${ commandLine.subcommand }")))
